Start of the biome registry system.
This commit contains the start to a biome registry system allowing us to temporarily add biomes to the overworld while we wait on Fabric API. The only issue is that because of our project structure access wideners are not functional. It would awesome if at some point @Wyn-Price or @coderbot could take a look at the project structure and try to get access wideners working so that we can move forward.